Add files via upload
To get these files to execute the correct data, you must download the files locally. In the CSV reader line, you have to include the path of the files. In my case, I have a directory that has these files downloaded in it as well as the other directories that include the data. The path must include the directory and the file name. Because the names are long and need to be copied, be sure to add .csv at the end of the path string.